Frontend_v2: Add edit challenge details feature to challenge settings tab(#2850)
* Move edit title, delete challenege and close participation option to settings

* Removed logger and edit start and end date feature added

* Minor UI Fix

* Edit challenge Image feature added

* Responsiveness Fix

* suggested changes
