+++ frontend_v2/src/app/components/challenge/challengesettings/challengesettings.component.scss
@@ -38,3 +38,12 @@
 .error-message {
     color: $highlight;
 }
+
+.challenge-detail {
+  margin-left: 25%;
+  padding: 8px;
+}
+
+.challenge-title {
+  color: rgba(0, 0, 0, 0.87);
+}
